+++ backend/src/device-comm/device-comm.module.ts
@@ -1,6 +1,6 @@
 import { Global, Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Device, AccessLog, Personnel, SyncHistory } from '../entities';
+import { Device, AccessLog, Personnel, SyncHistory, SystemSettings } from '../entities';
 import { ZktecoClientService } from './zkteco-client.service';
 import { DeviceManagerService } from './device-manager.service';
 import { SyncService } from './sync.service';
@@ -9,7 +9,7 @@ import { EmailModule } from '../email/email.module';
 @Global()
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Device, AccessLog, Personnel, SyncHistory]),
+    TypeOrmModule.forFeature([Device, AccessLog, Personnel, SyncHistory, SystemSettings]),
     EmailModule,
   ],
   providers: [ZktecoClientService, DeviceManagerService, SyncService],